fdfx

+++ ft_calloc.c
@@ -6,7 +6,7 @@
 /*   By: tmekhzou <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/11/07 01:58:42 by tmekhzou          #+#    #+#             */
-/*   Updated: 2023/11/07 17:55:13 by tmekhzou         ###   ########.fr       */
+/*   Updated: 2023/11/07 18:42:37 by tmekhzou         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 #include "libft.h"
@@ -18,7 +18,7 @@ void	*ft_calloc(size_t nmemb, size_t size)
 
 	if (!nmemb || !size)
 		return (malloc(0));
-	if (nmemb < 0 || size < 0)
+	if ((int)nmemb < 0 || (int)size < 0)
 		return (0);
 	total = size * nmemb;
 	if ((65535 / nmemb) >= size)
